add spec to capture trait bug
